chore: consolidate navigation components into a parent component
